Fix re-entrant lock attempt crash and make populate requests retry
This reduces some complexity around state management -- we don't need to care about whether we need to send a populate again due to state changes, because we can assume if one is outstanding that it'll be handled by the connection.

This mitigates a bug where our using the state change as a signal to start populate recursed due to populate, itself, causing a change to `.connecting`. State changes are now just a signal that a populate should happen at all. A second level of this was making the state change events happen in the callback queue, rather than whenever the state changed.

Fixes #17.